week-05/messages-svc-s15/app/main.py
import sys
from pathlib import Path
sys.path.insert(0, str(Path(__file__).parent))

from fastapi import FastAPI
import strawberry
from strawberry.fastapi import GraphQLRouter
from schema import Query, Mutation

schema = strawberry.Schema(query=Query, mutation=Mutation)
graphql_app = GraphQLRouter(schema)

app = FastAPI(title="Messages GraphQL Microservice")
app.include_router(graphql_app, prefix="/graphql")

@app.get("/health")
async def health_check():
    return {"status": "ok", "service": "messages-svc-s15"}