week-05/messages-svc-s15/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
strawberry-graphql==0.217.0
pydantic==2.4.2